Move "ui" application subclasses into better-named directories
Mainly organizing sidebar applications into a common directory
