+++ src/global.d.ts
@@ -7,9 +7,9 @@ import { CoinsPF2e } from "@item/physical/helpers";
 import { ActiveEffectPF2e } from "@module/active-effect";
 import { CompendiumBrowser, CompendiumBrowserSettings } from "@module/apps/compendium-browser";
 import { EffectsPanel } from "@module/apps/effects-panel";
+import { HotbarPF2e } from "@module/apps/hotbar";
 import { LicenseViewer } from "@module/apps/license-viewer";
-import { ActorDirectoryPF2e, ChatLogPF2e, CompendiumDirectoryPF2e, EncounterTrackerPF2e } from "@module/apps/ui";
-import { HotbarPF2e } from "@module/apps/ui/hotbar";
+import { ActorDirectoryPF2e, ChatLogPF2e, CompendiumDirectoryPF2e, EncounterTrackerPF2e } from "@module/apps/sidebar";
 import { WorldClock } from "@module/apps/world-clock";
 import { CanvasPF2e, EffectsCanvasGroupPF2e } from "@module/canvas";
 import { StatusEffects } from "@module/canvas/status-effects";
