+++ src/module/apps/sidebar/actor-directory.ts
