+++ src/module/apps/sidebar/chat-log.ts
