+++ src/module/apps/sidebar/compendium-directory.ts
