+++ src/module/apps/sidebar/index.ts
